pyproject.toml
[project]
name = "panel1"
version = "0.1.0"
description = "AI Coal Mine Reporting System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.49.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.0.0",
    "matplotlib>=3.9.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]
ocr = [
    "pytesseract>=0.3.13",
    "Pillow>=11.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
